@@ -7,7 +7,6 @@ import (
 	"encoding/hex"
 	"fmt"
 	"io"
-	"io/ioutil"
 	"net"
 	"net/http"
 	"net/http/httputil"
@@ -24,6 +23,7 @@ import (
 	"github.com/getlantern/flashlight/v7/email"
 	"github.com/getlantern/flashlight/v7/embeddedconfig"
 	"github.com/getlantern/flashlight/v7/geolookup"
+
 	"github.com/getlantern/lantern-client/internalsdk/common"
 )
 
@@ -135,7 +135,6 @@ func (cf *configurer) Configure(userID int, proToken string, refreshProxies bool
 		log.Errorf("Unable to configure fronting, sticking with embedded configuration: %v", err)
 	} else {
 		log.Debug("Refreshing geolookup")
-		geolookup.Refresh()
 
 		go func() {
 			cf.uc.Country = geolookup.GetCountry(1 * time.Minute)
@@ -208,7 +207,7 @@ func (cf *configurer) writeUserConfig() error {
 }
 
 func (cf *configurer) readUserConfig() (*UserConfig, error) {
-	bytes, err := ioutil.ReadFile(cf.fullPathTo(userConfigYaml))
+	bytes, err := os.ReadFile(cf.fullPathTo(userConfigYaml))
 	if err != nil {
 		return nil, errors.New("Unable to read userconfig.yaml: %v", err)
 	}
@@ -387,7 +386,7 @@ func (cf *configurer) doUpdateFromWeb(rt http.RoundTripper, name string, etag st
 		}
 	}()
 
-	bytes, err := ioutil.ReadAll(gzReader)
+	bytes, err := io.ReadAll(gzReader)
 	if err != nil {
 		return nil, "", errors.New("Unable to read response for %v: %v", name, err)
 	}
